examples: made the changes requested in the pr
