+++ examples/Cargo.toml
@@ -13,10 +13,10 @@ glow = { version = "0.13.0", optional = true }
 epoxy = { version = "0.1.0", optional = true }
 libloading = { version = "0.8.0", optional = true }
 im-rc = { version = "15", optional = true }
-async-channel = "2.1.1"
-tokio = { version = "1", features = ["full"] }
-reqwest = { version = "0.11", features = ["json"] }
-serde = { version = "1.0", features = ["derive"] }
+async-channel = { version = "2.1.1", optional = true }
+tokio = { version = "1", features = ["full"], optional = true }
+reqwest = { version = "0.11", features = ["json", "rustls-tls"], default-features = false, optional = true }
+serde = { version = "1.0", features = ["derive"], optional = true }
 
 [dependencies.gtk]
 path = "../gtk4"
@@ -36,6 +36,10 @@ femtovg-support = ["epoxy", "femtovg", "glow", "libloading"]
 glium-support = ["glium", "epoxy", "libloading"]
 v4_12 = ["gtk/v4_12"]
 v4_14 = ["gtk/v4_14"]
+tokio-support = ["tokio"]
+async-channel-support = ["async-channel"]
+reqwest-support = ["reqwest"]
+serde-support = ["serde"]
 
 [[bin]]
 name = "basics"
@@ -193,13 +197,14 @@ name = "text_viewer"
 path = "text_viewer/main.rs"
 
 [[bin]]
+name = "tokio_async_request"
+path = "tokio_async_request/main.rs"
+required-features = ["tokio-support", "async-channel-support", "reqwest-support", "serde-support"]
+
+[[bin]]
 name = "video_player"
 path = "video_player/main.rs"
 
 [[bin]]
 name = "virtual_methods"
 path = "virtual_methods/main.rs"
-
-[[bin]]
-name = "request_no_blocking_main_thread"
-path = "request_no_blocking_main_thread/main.rs"
