+++ examples/tokio_async_request/README.md
@@ -1,0 +1,5 @@
+# Tokio async request
+
+This example shows an infinite list of pokemon names, whenever it reaches the end, it searches for new names and fills the list without blocking the main thread of the program.
+
+![Screenshot](screenshot.gif)
